06_claude/pyproject.toml
[project]
name = "06-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.88.0",
    "python-dotenv>=1.2.2",
]
